Fixed crash on shutdown, removing base class for ui
