+++ src/ReplayHandler.cpp
@@ -8,26 +8,22 @@
 #include <rtt/transports/corba/CorbaDispatcher.hpp>
 #include <rtt/typelib/TypelibMarshallerBase.hpp>
 #include <typelib/typedisplay.hh>
+#include <orocos_cpp/orocos_cpp.hpp>
+#include "FileLoader.hpp"
 
 
 void ReplayHandler::loadStreams(int argc, char** argv, MATCH_MODE mode)
 {
-    RTT::corba::TaskContextServer::InitOrb(argc, argv);    
-    
-    // load basic typekits
-    orocos_cpp::PluginHelper::loadTypekitAndTransports("orocos");
-    
-    multiIndex = new pocolog_cpp::MultiFileIndex();    
-    
-    orocos_cpp::TypeRegistry reg;
-    reg.loadTypeRegistries();
-    RTT::types::TypeInfoRepository::shared_ptr ti = RTT::types::TypeInfoRepository::Instance();
+    orocos_cpp::OrocosCppConfig config;
+    config.load_all_packages = true;
+    orocos_cpp::OrocosCpp orocos;
+    orocos.initialize(config);
 
     std::vector<std::regex> regExps;
     std::map<std::string, std::string> logfiles2Prefix;
-    std::vector<std::string> fileNames = parseFilenames(argc, argv, regExps, logfiles2Prefix);
-
-    multiIndex->registerStreamCheck([&](pocolog_cpp::Stream *st){
+    std::vector<std::string> fileNames = FileLoader::parseFileNames(argc, argv, regExps, logfiles2Prefix);
+    
+    multiIndex.registerStreamCheck([&](pocolog_cpp::Stream *st){
         std::cout << "Checking " << st->getName() << std::endl;
         pocolog_cpp::InputDataStream *dataStream = dynamic_cast<pocolog_cpp::InputDataStream *>(st);
         if(!dataStream)
@@ -35,48 +31,13 @@ void ReplayHandler::loadStreams(int argc, char** argv, MATCH_MODE mode)
             return false;
         }
         
-        bool matches;
-        switch(mode)
-        {
-            case REGEX:
-                matches = regExps.empty();
-                for(std::regex exp : regExps)
-                {
-                    if(std::regex_search(dataStream->getName(), exp))
-                        matches = true;
-                }
-                break;
-            case WHITELIST:
-                matches = whiteList.find(dataStream->getName()) != whiteList.end();
-                break;
-        }
         
-        if(!matches)
-        {
-            std::cout << "skipping non-whitelisted stream " << dataStream->getName() << std::endl;
-            return false;
-        }
+        std::string modelName = getTaskName(dataStream);
         
-        std::string typestr = dataStream->getType()->getName();
         
-        std::string tkName;
-        if(!reg.getTypekitDefiningType(typestr, tkName))
+        if(!orocos.loadAllTypekitsForModel(modelName))
         {
-            std::cerr << "cannot find " << typestr << " in the type info repository" << std::endl;
-            return false;
-        }
-
-        
-        if(!orocos_cpp::PluginHelper::loadTypekitAndTransports(tkName))
-        {
-            return false;
-        }
-
-   
-        RTT::types::TypeInfo* type = ti->type(dataStream->getCXXType());
-        if (! type)
-        {
-            std::cerr << "2 cannot find " << typestr << " in the type info repository" << std::endl;
+            std::cerr << "cannot find " << modelName << " in the type info repository" << std::endl;
             return false;
         }
         
@@ -85,26 +46,16 @@ void ReplayHandler::loadStreams(int argc, char** argv, MATCH_MODE mode)
     }
     );
 
-    multiIndex->createIndex(fileNames);
-    streamToTask.resize(multiIndex->getAllStreams().size());
+    multiIndex.createIndex(fileNames);
+    streamToTask.resize(multiIndex.getAllStreams().size());
 
-    for(pocolog_cpp::Stream *st : multiIndex->getAllStreams())
+    for(pocolog_cpp::Stream *st : multiIndex.getAllStreams())
     {
         pocolog_cpp::InputDataStream *inputSt = dynamic_cast<pocolog_cpp::InputDataStream *>(st);
         if(!inputSt)
             continue;       
-
-        std::string taskName = st->getName();
         
-        // filter all '/'
-        size_t pos = taskName.find('/');
-        if(taskName.size() && pos != std::string::npos)
-        {
-            taskName = taskName.substr(pos + 1, taskName.size());
-        }
-        
-        // remove port name
-        taskName = taskName.substr(0, taskName.find_last_of('.'));
+        std::string taskName = getTaskName(inputSt);
         
         //add prefix if active
         if(logfiles2Prefix.find(inputSt->getFileStream().getFileName()) != logfiles2Prefix.end())
@@ -126,14 +77,14 @@ void ReplayHandler::loadStreams(int argc, char** argv, MATCH_MODE mode)
         }
         task->addStream(*inputSt);
 
-        size_t gIdx = multiIndex->getGlobalStreamIdx(st); 
+        size_t gIdx = multiIndex.getGlobalStreamIdx(st); 
         if(gIdx > streamToTask.size())
             throw std::runtime_error("Mixup detected");
 
         streamToTask[gIdx] = task;
     }
 
-    valid = !multiIndex->getAllStreams().empty() && !fileNames.empty() && multiIndex->getSize() > 0;
+    valid = !multiIndex.getAllStreams().empty() && !fileNames.empty() && multiIndex.getSize() > 0;
     if(!valid)
     {
         std::cerr << "empty streams loaded" << std::endl;
@@ -146,105 +97,37 @@ void ReplayHandler::loadStreams(int argc, char** argv, MATCH_MODE mode)
 
 }
 
-
-std::vector<std::string> ReplayHandler::parseFilenames(int argc, char* argv[], std::vector<std::regex>& regExps, std::map<std::string, std::string>& logfiles2Prefix)
+std::string ReplayHandler::getTaskName(pocolog_cpp::InputDataStream* stream)
 {
-    std::vector<std::string> filenames;
-    
-    for(int i = 1; i < argc; i++)
+    std::string taskName = stream->getName();
+        
+    // filter all '/'
+    size_t pos = taskName.find('/');
+    if(taskName.size() && pos != std::string::npos)
     {
-        std::string argv2String(argv[i]);
-        std::string whiteList = "--white-list=";
-        std::size_t pos = argv2String.find(whiteList);
-        
-        if(pos != std::string::npos) 
-        {
-            std::string params = argv2String.substr(whiteList.length(), argv2String.length());
-            boost::split(regExps, params, boost::is_any_of(","));
-            continue;
-        }
-    
-        std::string prefix = "--prefix=";
-        pos = argv2String.find(prefix);
-    
-        if(pos != std::string::npos) 
-        {
-            std::string params = argv2String.substr(prefix.length(), argv2String.length());
-            std::vector<std::string> prefixTuples;
-            boost::split(prefixTuples, params, boost::is_any_of(","));
-            
-            for(const std::string pair : prefixTuples)
-            {
-                std::size_t sep = pair.find(":");
-                if(sep != std::string::npos)
-                {
-                    std::string logfile = pair.substr(0, sep);
-                    std::string prefix = pair.substr(sep + 1, pair.length());
-                    
-                    std::cout << "prefixing: " << logfile << " -> " << prefix << std::endl;
-                    logfiles2Prefix.insert(std::make_pair(logfile, prefix));
-                }
-            }
-            
-            continue;
-        }
-        
-        struct stat file_stat;
-        if(stat(argv[i], &file_stat) == -1)
-        {
-            std::cerr << "stat error: couldn't open folder" << std::endl;
-            exit(EXIT_FAILURE);
-        }
-        
-        if(S_ISDIR(file_stat.st_mode))
-        {
-            DIR *dir = opendir(argv[i]);
-            if(!dir) 
-            {
-                std::cerr << "directory opening error" << std::endl;
-                exit(EXIT_FAILURE);   
-            }
-            else
-            {
-                dirent *entry;
-                std::string dir_path = std::string(argv[i]);
-                if(dir_path.back() != '/')
-                    dir_path.append("/");
-                
-                while((entry = readdir(dir)) != 0)
-                {
-                    std::string filename = entry->d_name;
-                    if(filename.substr(filename.find_last_of(".") + 1) == "log")
-                    {
-                        if(filename == "orocos.log")
-                            continue;
-                        
-                        filenames.push_back(std::string(dir_path).append(filename));
-                    }
-                }
-            }
-            closedir(dir);
-        }
-        else if(S_ISREG(file_stat.st_mode))
-        {
-            filenames.push_back(argv[i]);
-        }
+        taskName = taskName.substr(pos + 1, taskName.size());
     }
     
-    return filenames;
+    // remove port name
+    taskName = taskName.substr(0, taskName.find_last_of('.'));
+    
+    return taskName;
 }
-
-
 
 ReplayHandler::~ReplayHandler()
 {       
     RTT::corba::CorbaDispatcher::ReleaseAll();
     RTT::corba::TaskContextServer::CleanupServers();
-       
+//     playing = false;
+//     running = false;
+//     cond.notify_all();
+    replayThread.join();
+    
+    
     for(std::map<std::string, LogTask*>::iterator it = logTasks.begin(); it != logTasks.end(); it++)
         delete it->second;
         
-    delete multiIndex;    
+    //delete multiIndex;    
 }
 
 
@@ -255,9 +138,9 @@ void ReplayHandler::init()
     curIndex = 0;
     finished = false;
     playing = false;
-    maxIndex = multiIndex->getSize() - 1;
+    maxIndex = multiIndex.getSize() - 1;
     replaySample(curIndex, true);       
-    boost::thread(boost::bind(&ReplayHandler::replaySamples, boost::ref(*this)));
+    replayThread = std::thread(std::bind(&ReplayHandler::replaySamples, this));
 }
 
 const base::Time ReplayHandler::getMinIdxTimeStamp()
@@ -276,9 +159,9 @@ bool ReplayHandler::replaySample(size_t index, bool dryRun)
 {
     try 
     {
-        size_t globalStreamIndex = multiIndex->getGlobalStreamIdx(index);
-        pocolog_cpp::InputDataStream *inputSt = dynamic_cast<pocolog_cpp::InputDataStream *>(multiIndex->getSampleStream(index));
-        if (dryRun || (!dryRun && streamToTask[globalStreamIndex]->replaySample(*inputSt, multiIndex->getPosInStream(index))))
+        size_t globalStreamIndex = multiIndex.getGlobalStreamIdx(index);
+        pocolog_cpp::InputDataStream *inputSt = dynamic_cast<pocolog_cpp::InputDataStream *>(multiIndex.getSampleStream(index));
+        if (dryRun || (!dryRun && streamToTask[globalStreamIndex]->replaySample(*inputSt, multiIndex.getPosInStream(index))))
         {
             curSamplePortName = inputSt->getName();
             curTimeStamp = getTimeStamp(index).toString();
@@ -309,92 +192,92 @@ bool ReplayHandler::checkSampleIdx()
 
 void ReplayHandler::replaySamples()
 {   
-    boost::unique_lock<boost::mutex> lock(mut);
+//     boost::unique_lock<boost::mutex> lock(mut);
+//     running = true;
+//     restartReplay = true;
+//     
+//     base::Time curStamp;
+//     base::Time toSleep;
+// 
+//     base::Time systemPlayStartTime;
+//     base::Time logPlayStartTime;
+//     curStamp = getTimeStamp(curIndex);
     
-    restartReplay = true;
-    
-    base::Time curStamp;
-    base::Time toSleep;
-
-    base::Time systemPlayStartTime;
-    base::Time logPlayStartTime;
-    curStamp = getTimeStamp(curIndex);
-    
-    while(1)
-    {
-        while(!playing)
-        {
-            cond.wait(lock);
-            restartReplay = true;
-        }
-        
-        varMut.lock();
-        
-        if(checkSampleIdx())
-            continue;
-        
-        
-        if(restartReplay)
-        {
-            systemPlayStartTime = base::Time::now();
-            
-            //expensive call
-            curStamp = getTimeStamp(curIndex);
-            logPlayStartTime = curStamp;
-            restartReplay = false;
-        }
-        
-        //TODO check if chronological ordering is right
-        //TODO if logging for port is not selected, skip cur index
-        // (allow higher replayFactor)
-        if (!replaySample(curIndex++))
-        {
-            varMut.unlock();
-            continue;
-        }
-        
-        if(checkSampleIdx())
-            continue;
-        
-        replaySample(curIndex, true);
-        curStamp = getTimeStamp(curIndex);    
-
-        //hm, also expensive, is there a way to reduce this ?
-        base::Time curTime = base::Time::now();
-        
-        //hm, factor... should it not be * replayFactor then ?
-        int64_t tSinceStart = curStamp.microseconds - logPlayStartTime.microseconds;
-        tSinceStart = std::max<int64_t>(0, tSinceStart);
-        base::Time logTimeSinceStart = base::Time::fromMicroseconds(tSinceStart / replayFactor);
-        base::Time systemTimeSinceStart = (curTime - systemPlayStartTime);
-        toSleep = logTimeSinceStart - systemTimeSinceStart;
-
-        varMut.unlock();
-        
-        if(!playing)
-            continue;
-        
-        if(toSleep.microseconds > 0)
-        {
-            factorChangeCond.timed_wait(lock, boost::posix_time::microseconds(toSleep.microseconds));
-            currentSpeed = replayFactor;
-        }
-        else if(toSleep.microseconds == 0)
-        {
-            currentSpeed = replayFactor;
-        }
-        else // tosleep < 0
-        {
-            currentSpeed = logTimeSinceStart.toSeconds() / systemTimeSinceStart.toSeconds();   
-        }      
-    }
+//     while(running)
+//     {
+//         while(!playing)
+//         {
+//             cond.wait(lock);
+//             restartReplay = true;
+//         }
+//         
+//         varMut.lock();
+//         
+//         if(checkSampleIdx())
+//             continue;
+//         
+//         
+//         if(restartReplay)
+//         {
+//             systemPlayStartTime = base::Time::now();
+//             
+//             //expensive call
+//             curStamp = getTimeStamp(curIndex);
+//             logPlayStartTime = curStamp;
+//             restartReplay = false;
+//         }
+//         
+//         //TODO check if chronological ordering is right
+//         //TODO if logging for port is not selected, skip cur index
+//         // (allow higher replayFactor)
+//         if (!replaySample(curIndex++))
+//         {
+//             varMut.unlock();
+//             continue;
+//         }
+//         
+//         if(checkSampleIdx())
+//             continue;
+//         
+//         replaySample(curIndex, true);
+//         curStamp = getTimeStamp(curIndex);    
+// 
+//         //hm, also expensive, is there a way to reduce this ?
+//         base::Time curTime = base::Time::now();
+//         
+//         //hm, factor... should it not be * replayFactor then ?
+//         int64_t tSinceStart = curStamp.microseconds - logPlayStartTime.microseconds;
+//         tSinceStart = std::max<int64_t>(0, tSinceStart);
+//         base::Time logTimeSinceStart = base::Time::fromMicroseconds(tSinceStart / replayFactor);
+//         base::Time systemTimeSinceStart = (curTime - systemPlayStartTime);
+//         toSleep = logTimeSinceStart - systemTimeSinceStart;
+// 
+//         varMut.unlock();
+//         
+//         if(!playing)
+//             continue;
+//         
+//         if(toSleep.microseconds > 0)
+//         {
+//             factorChangeCond.timed_wait(lock, boost::posix_time::microseconds(toSleep.microseconds));
+//             currentSpeed = replayFactor;
+//         }
+//         else if(toSleep.microseconds == 0)
+//         {
+//             currentSpeed = replayFactor;
+//         }
+//         else // tosleep < 0
+//         {
+//             currentSpeed = logTimeSinceStart.toSeconds() / systemTimeSinceStart.toSeconds();   
+//         }      
+//     }
 
 }
 
 const base::Time ReplayHandler::getTimeStamp(size_t globalIndex)
 {    
-    pocolog_cpp::Index &idx = multiIndex->getSampleStream(globalIndex)->getFileIndex();
-    return idx.getSampleTime(multiIndex->getPosInStream(globalIndex));
+    pocolog_cpp::Index &idx = multiIndex.getSampleStream(globalIndex)->getFileIndex();
+    return idx.getSampleTime(multiIndex.getPosInStream(globalIndex));
 }
 
 void ReplayHandler::stop()
